Split DepositTest file
